+++ Entities/GroundedCharacter.h
@@ -33,7 +33,7 @@ protected:
 
 public:
     GroundedCharacter(const char* fileName, double xStartPos = 0, double yStartPos = 0, double xVel = 0, 
-        double yVel = 0, int colliderWidth = 0, int colliderHeight = 0, int hitPoints = 0, int spriteSheetCount = 1);
+        double yVel = 0, double colliderWidth = 0, double colliderHeight = 0, int hitPoints = 0, int spriteSheetCount = 1);
     virtual ~GroundedCharacter();
 
     Movement getMovement() const { return m_movement; }
